+++ crates/hyperqueue/src/client/output/cli.rs
@@ -1671,12 +1671,17 @@ fn resources_summary(resources: &ResourceDescriptor, multiline: bool) -> String 
         None
     };
 
+    let mut coupled = tako::Set::new();
+    for w in &resources.coupling.weights {
+        coupled.insert(w.resource1_idx);
+        coupled.insert(w.resource2_idx);
+    }
     let mut result = String::new();
     let mut first = true;
-    for descriptor in &resources.resources {
+    for (i, descriptor) in resources.resources.iter().enumerate() {
         write!(
             result,
-            "{}{}{} {}",
+            "{}{}{} {}{}",
             if first {
                 ""
             } else if multiline {
@@ -1686,17 +1691,14 @@ fn resources_summary(resources: &ResourceDescriptor, multiline: bool) -> String 
             },
             &descriptor.name,
             if multiline { ":" } else { "" },
-            special_format(descriptor).unwrap_or_else(|| resource_summary_kind(&descriptor.kind))
+            special_format(descriptor).unwrap_or_else(|| resource_summary_kind(&descriptor.kind)),
+            if coupled.contains(&(i as u8)) {
+                " [coupled]"
+            } else {
+                ""
+            }
         )
         .unwrap();
-        if multiline
-            && resources
-                .coupling
-                .as_ref()
-                .is_some_and(|c| c.names.contains(&descriptor.name))
-        {
-            result.push_str(" [coupled]");
-        }
         first = false;
     }
     result
@@ -1762,7 +1764,7 @@ mod tests {
                 name: "cpus".into(),
                 kind: ResourceDescriptorKind::simple_indices(1),
             }],
-            None,
+            Default::default(),
         );
         assert_eq!(resources_summary(&d, false), "cpus 1");
 
@@ -1771,7 +1773,7 @@ mod tests {
                 name: "cpus".into(),
                 kind: ResourceDescriptorKind::simple_indices(5),
             }],
-            None,
+            Default::default(),
         );
         assert_eq!(resources_summary(&d, true), "cpus: 5");
 
@@ -1804,7 +1806,7 @@ mod tests {
                     ]),
                 },
             ],
-            None,
+            Default::default(),
         );
         assert_eq!(
             resources_summary(&d, true),
@@ -1823,7 +1825,7 @@ mod tests {
                 name: "cpus".into(),
                 kind: ResourceDescriptorKind::simple_indices(1),
             }],
-            None,
+            Default::default(),
         );
         assert_eq!(resources_full_describe(&d), "cpus: [0]");
 
@@ -1845,7 +1847,7 @@ mod tests {
                     kind: res_kind_sum(1234),
                 },
             ],
-            None,
+            Default::default(),
         );
         assert_eq!(
             resources_full_describe(&d),
@@ -1860,7 +1862,7 @@ mod tests {
                 name: MEM_RESOURCE_NAME.into(),
                 kind: res_kind_sum(4 * 1024 + 123),
             }],
-            None,
+            Default::default(),
         );
         assert_eq!(resources_summary(&d, false), "mem 4.12 GiB");
     }