+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 * Client notification mechanism. A task can generate a notification,
   which the process that submitted the task can capture.
 * `hq journal report` that exports a statistics into an HTML report
+* Resource coupling is now stabilized. It is extended to freely set arbitrary weights between groups.
 * Server scheduler more respects priorities
 * Server scheduler improved when resource variants are used
 * Extended events:
@@ -26,6 +27,7 @@
 * `hq job progress` stopped using periodic polling.
   In the current version, the client passively waits for events.
   It saves some resources as the server is not queried every second.
+
 
 ## v0.24.0